requirement.txt
numpy==1.21.2
pandas==1.3.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.5
decorator==5.1.0
et-xmlfile==1.1.0
idna==3.2
ipython==7.27.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
numpy==1.21.2
openpyxl==3.0.8
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
tqdm==4.62.2
traitlets==5.1.0
urllib3==1.26.6
wcwidth==0.2.5
backcall==0.2.0
black==21.9b0
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.0.1
decorator==5.1.0
et-xmlfile==1.1.0
idna==3.2
ipython==7.27.0
jedi==0.18.0
Jinja2==3.0.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
numpy==1.21.2
openpyxl==3.0.8
pandas==1.3.3
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
platformdirs==2.3.0
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
python-dateutil==2.8.2
pytz==2021.1
regex==2021.8.28
requests==2.26.0
six==1.16.0
tomli==1.2.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
wcwidth==0.2.5
